+++ NEWS.md
@@ -16,7 +16,7 @@
 * `confint_betabinom()`: Methods `"kaplan"` and `"nelson"` of `estimate_cdf()` can be used for beta-binomial confidence bounds. 
 
 ## Minor Improvements and bug fixes
-* The object returned by `dist_delay()` know has class `wt_delay_estimation` or `wt_delay_estimation_list`. 
+* The object returned by `dist_delay()` now has class `wt_delay_estimation` or `wt_delay_estimation_list`. 
 * Fixed bug in `plot_conf()`: Wrong confidence bounds were displayed for `direction = "x"` (#181). 
 * Fixed bug in `plot_conf()`: `plot_method = "ggplot2"` and exactly one method in `estimate_cdf()` resulted in an error (#182).
 * Fixed bug in `reliability_data()`: Using `!!` syntax with arguments `x` and `status` resulted in an error.